+++ compiler/src/org.graalvm.compiler.truffle.pelang/src/org/graalvm/compiler/truffle/pelang/array/PELangNewArrayNode.java
@@ -20,7 +20,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package org.graalvm.compiler.truffle.pelang.expr;
+package org.graalvm.compiler.truffle.pelang.array;
 
 import java.lang.reflect.Array;
