+++ compiler/src/org.graalvm.compiler.truffle.pelang/src/org/graalvm/compiler/truffle/pelang/util/PELangBuilder.java
@@ -33,42 +33,42 @@ import org.graalvm.compiler.truffle.pelang.PELangPrintNode;
 import org.graalvm.compiler.truffle.pelang.PELangRootNode;
 import org.graalvm.compiler.truffle.pelang.PELangState;
 import org.graalvm.compiler.truffle.pelang.PELangStatementNode;
+import org.graalvm.compiler.truffle.pelang.array.PELangArrayType;
+import org.graalvm.compiler.truffle.pelang.array.PELangNewArrayNode;
+import org.graalvm.compiler.truffle.pelang.array.PELangReadArrayNode;
+import org.graalvm.compiler.truffle.pelang.array.PELangWriteArrayNode;
 import org.graalvm.compiler.truffle.pelang.bcf.PELangBasicBlockDispatchNode;
 import org.graalvm.compiler.truffle.pelang.bcf.PELangBasicBlockNode;
 import org.graalvm.compiler.truffle.pelang.bcf.PELangDoubleSuccessorNode;
 import org.graalvm.compiler.truffle.pelang.bcf.PELangMultiSuccessorNode;
 import org.graalvm.compiler.truffle.pelang.bcf.PELangSingleSuccessorNode;
+import org.graalvm.compiler.truffle.pelang.call.PELangInvokeNode;
 import org.graalvm.compiler.truffle.pelang.expr.PELangAddNode;
-import org.graalvm.compiler.truffle.pelang.expr.PELangArrayType;
 import org.graalvm.compiler.truffle.pelang.expr.PELangDivNode;
 import org.graalvm.compiler.truffle.pelang.expr.PELangEqualsNode;
 import org.graalvm.compiler.truffle.pelang.expr.PELangGreaterThanNode;
-import org.graalvm.compiler.truffle.pelang.expr.PELangInvokeNode;
 import org.graalvm.compiler.truffle.pelang.expr.PELangLeftShiftNode;
 import org.graalvm.compiler.truffle.pelang.expr.PELangLessThanNode;
-import org.graalvm.compiler.truffle.pelang.expr.PELangLiteralArrayNode;
-import org.graalvm.compiler.truffle.pelang.expr.PELangLiteralFunctionNode;
-import org.graalvm.compiler.truffle.pelang.expr.PELangLiteralLongNode;
-import org.graalvm.compiler.truffle.pelang.expr.PELangLiteralNullNode;
-import org.graalvm.compiler.truffle.pelang.expr.PELangLiteralStringNode;
 import org.graalvm.compiler.truffle.pelang.expr.PELangMinusNode;
-import org.graalvm.compiler.truffle.pelang.expr.PELangNewArrayNode;
-import org.graalvm.compiler.truffle.pelang.expr.PELangNewObjectNode;
 import org.graalvm.compiler.truffle.pelang.expr.PELangNotNode;
-import org.graalvm.compiler.truffle.pelang.expr.PELangReadArgumentNode;
-import org.graalvm.compiler.truffle.pelang.expr.PELangReadArrayNode;
-import org.graalvm.compiler.truffle.pelang.expr.PELangWriteArrayNode;
+import org.graalvm.compiler.truffle.pelang.lit.PELangLiteralArrayNode;
+import org.graalvm.compiler.truffle.pelang.lit.PELangLiteralFunctionNode;
+import org.graalvm.compiler.truffle.pelang.lit.PELangLiteralLongNode;
+import org.graalvm.compiler.truffle.pelang.lit.PELangLiteralNullNode;
+import org.graalvm.compiler.truffle.pelang.lit.PELangLiteralStringNode;
 import org.graalvm.compiler.truffle.pelang.ncf.PELangBlockNode;
 import org.graalvm.compiler.truffle.pelang.ncf.PELangIfNode;
 import org.graalvm.compiler.truffle.pelang.ncf.PELangReturnNode;
 import org.graalvm.compiler.truffle.pelang.ncf.PELangSwitchNode;
 import org.graalvm.compiler.truffle.pelang.ncf.PELangWhileNode;
+import org.graalvm.compiler.truffle.pelang.obj.PELangNewObjectNode;
 import org.graalvm.compiler.truffle.pelang.obj.PELangPropertyReadNode;
 import org.graalvm.compiler.truffle.pelang.obj.PELangPropertyWriteNode;
 import org.graalvm.compiler.truffle.pelang.var.PELangGlobalReadNode;
 import org.graalvm.compiler.truffle.pelang.var.PELangGlobalWriteNode;
 import org.graalvm.compiler.truffle.pelang.var.PELangLocalReadNode;
 import org.graalvm.compiler.truffle.pelang.var.PELangLocalWriteNode;
+import org.graalvm.compiler.truffle.pelang.var.PELangReadArgumentNode;
 
 import com.oracle.truffle.api.Truffle;
 import com.oracle.truffle.api.frame.FrameDescriptor;
